# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Passport Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "bosavras@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/user-login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/user-register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mobile_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/news": {
            "get": {
                "summary": "Return news",
                "description": "Get news",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\NewsController::getAll",
                "responses": {
                    "200": {
                        "description": "Array of new",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "title": {
                                                "description": "Title",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "content": {
                                                "description": "Content",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "createAt": {
                                                "description": "Create at",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "updateAt": {
                                                "description": "Update At",
                                                "type": "integer",
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update news by id",
                "description": "Update news by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\NewsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Update new body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Tile",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "content": {
                                        "description": "Content",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "new",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "title": {
                                            "description": "Title",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "content": {
                                            "description": "Content",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "createAt": {
                                            "description": "Create at",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "updateAt": {
                                            "description": "Update At",
                                            "type": "integer",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exceptions"
                    }
                }
            },
            "post": {
                "summary": "Create news",
                "description": "Create news",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\NewsController::create",
                "requestBody": {
                    "description": "New body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Tile",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "content": {
                                        "description": "Content",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation exceptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "title": {
                                            "description": "Title",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "content": {
                                            "description": "Content",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "createAt": {
                                            "description": "Create at",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "updateAt": {
                                            "description": "Update At",
                                            "type": "integer",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/news/{id}": {
            "get": {
                "summary": "Return new by id",
                "description": "Get new by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\NewsController::getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "title": {
                                            "description": "Title",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "content": {
                                            "description": "Content",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "createAt": {
                                            "description": "Create at",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "updateAt": {
                                            "description": "Update At",
                                            "type": "integer",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete news by id",
                "description": "Delete news by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\NewsController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "news id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "401": {
                "description": "Unauthorized"
            },
            "404": {
                "description": "Unauthorized"
            }
        }
    },
    "tags": [
        {
            "name": "Event"
        },
        {
            "name": "News"
        }
    ],
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}